+++ rotkehlchen/chain/evm/decoding/paraswap/decoder.py
@@ -1,10 +1,9 @@
 import logging
-from collections.abc import Callable
-from typing import TYPE_CHECKING, Any
+from abc import ABC
+from typing import TYPE_CHECKING
 
 from rotkehlchen.accounting.structures.balance import Balance
 from rotkehlchen.assets.asset import CryptoAsset, EvmToken
-from rotkehlchen.chain.ethereum.modules.uniswap.v2.constants import SWAP_SIGNATURE
 from rotkehlchen.chain.ethereum.utils import asset_normalized_value
 from rotkehlchen.chain.evm.decoding.constants import ERC20_OR_ERC721_TRANSFER
 from rotkehlchen.chain.evm.decoding.interfaces import DecoderInterface
@@ -14,7 +13,6 @@ from rotkehlchen.chain.evm.decoding.structures import (
     DecodingOutput,
 )
 from rotkehlchen.chain.evm.decoding.types import CounterpartyDetails
-from rotkehlchen.chain.evm.decoding.uniswap.v3.constants import DIRECT_SWAP_SIGNATURE
 from rotkehlchen.chain.evm.decoding.utils import maybe_reshuffle_events
 from rotkehlchen.chain.evm.transactions import EvmTransactions
 from rotkehlchen.errors.misc import RemoteError
@@ -23,15 +21,7 @@ from rotkehlchen.logging import RotkehlchenLogsAdapter
 from rotkehlchen.types import ChecksumEvmAddress
 from rotkehlchen.utils.misc import bytes_to_address
 
-from .constants import (
-    BUY_ON_UNISWAP_V2_FORK,
-    BUY_SIGNATURE,
-    CPT_PARASWAP,
-    SWAP_ON_UNISWAP_V2_FACTORY,
-    SWAP_ON_UNISWAP_V2_FORK,
-    SWAP_ON_UNISWAP_V2_FORK_WITH_PERMIT,
-    SWAP_SIGNATURE as PARASWAP_SWAP_SIGNATURE,
-)
+from .constants import CPT_PARASWAP
 
 if TYPE_CHECKING:
     from rotkehlchen.chain.evm.decoding.base import BaseDecoderTools
@@ -43,7 +33,8 @@ logger = logging.getLogger(__name__)
 log = RotkehlchenLogsAdapter(logger)
 
 
-class ParaswapCommonDecoder(DecoderInterface):
+class ParaswapCommonDecoder(DecoderInterface, ABC):
+
     def __init__(
             self,
             evm_inquirer: 'EvmNodeInquirer',
@@ -60,11 +51,14 @@ class ParaswapCommonDecoder(DecoderInterface):
     def _decode_swap(
             self,
             context: DecoderContext,
-            receiver: ChecksumEvmAddress,
             sender: ChecksumEvmAddress,
+            receiver: ChecksumEvmAddress | None = None,
     ) -> DecodingOutput:
-        """This function is used to decode the swap done by paraswap."""
-        if not self.base.any_tracked((sender, receiver)):
+        """This function is used to decode the swap done by paraswap.
+        In v6 swaps, receiver is unavailable and is ignored,
+        since sender should always be a tracked address.
+        """
+        if not self.base.any_tracked((sender,) if receiver is None else (sender, receiver)):
             return DEFAULT_DECODING_OUTPUT
 
         out_event: EvmEvent | None = None
@@ -157,7 +151,7 @@ class ParaswapCommonDecoder(DecoderInterface):
             fee_event = self.base.make_event_from_transaction(
                 transaction=context.transaction,
                 tx_log=context.tx_log,
-                event_type=HistoryEventType.SPEND,
+                event_type=HistoryEventType.TRADE,
                 event_subtype=HistoryEventSubType.FEE,
                 asset=fee_asset,
                 balance=Balance(amount=fee_amount),
@@ -174,49 +168,7 @@ class ParaswapCommonDecoder(DecoderInterface):
         )
         return DEFAULT_DECODING_OUTPUT
 
-    def _decode_paraswap_swap(self, context: DecoderContext) -> DecodingOutput:
-        """This decodes the following types of trades:
-        - Simple Buy
-        - Simple Swap
-        - Multi Swap
-        - Mega Swap
-        - Direct Swap on Uniswap V3
-        - Direct Swap on Curve V1 and V2
-        - Direct Swap on Balancer V2"""
-        if context.tx_log.topics[0] not in {PARASWAP_SWAP_SIGNATURE, BUY_SIGNATURE, DIRECT_SWAP_SIGNATURE}:  # noqa: E501
-            return DEFAULT_DECODING_OUTPUT
-
-        return self._decode_swap(
-            context=context,
-            receiver=bytes_to_address(context.tx_log.topics[1]),
-            sender=bytes_to_address(context.tx_log.data[96:128]),
-        )
-
-    def _decode_uniswap_v2_swap(self, context: DecoderContext) -> DecodingOutput:
-        """This decodes swaps done directly on Uniswap V2 pools"""
-        return self._decode_swap(
-            context=context,
-            receiver=bytes_to_address(context.tx_log.topics[2]),
-            sender=context.transaction.from_address,
-        )
-
     # -- DecoderInterface methods
-
-    def addresses_to_decoders(self) -> dict[ChecksumEvmAddress, tuple[Any, ...]]:
-        return {
-            self.router_address: (self._decode_paraswap_swap,),
-        }
-
-    def decoding_by_input_data(self) -> dict[bytes, dict[bytes, Callable]]:
-        return {
-            method_id: {SWAP_SIGNATURE: self._decode_uniswap_v2_swap}
-            for method_id in (
-                SWAP_ON_UNISWAP_V2_FORK,
-                SWAP_ON_UNISWAP_V2_FACTORY,
-                SWAP_ON_UNISWAP_V2_FORK_WITH_PERMIT,
-                BUY_ON_UNISWAP_V2_FORK,
-            )
-        }
 
     @staticmethod
     def counterparties() -> tuple[CounterpartyDetails, ...]: