+++ rotkehlchen/chain/optimism/modules/paraswap/v6/decoder.py
@@ -1,0 +1,5 @@
+from rotkehlchen.chain.evm.decoding.paraswap.v6.decoder import Paraswapv6CommonDecoder
+
+
+class Paraswapv6Decoder(Paraswapv6CommonDecoder):
+    ...