+++ docs/changelog.rst
@@ -2,6 +2,7 @@
 Changelog
 =========
 
+* :feature:`9083` rotki will now correctly decode ParaSwap v6 swaps in all supported EVM chains.
 * :feature:`9120` rotki will now correctly decode Firebird Finance swaps in all supported EVM chains.
 * :feature:`7254` rotki will now correctly decode OpenOcean swaps in all supported EVM chains.
 * :bug:`-` Bitmex balances will now be queried correctly.